requirements.txt
# Core dependencies
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# Flask API
flask>=3.0.0
flask-cors>=4.0.0
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.11.1

# Optional: for future enhancements
# faiss-cpu>=1.7.4  # For vector search (memory)
# chromadb>=0.4.0  # Alternative vector DB
# redis>=5.0.0  # For distributed rate limiting

